Refactor out code to common class
