--- src/test/java/com/palantir/docker/compose/execution/Retryer.java
@@ -1,11 +1,0 @@
-package com.palantir.docker.compose.execution;
-
-public class Retryer {
-    private final int attempts;
-
-    public Retryer(int attempts) {
-        this.attempts = attempts;
-    }
-
-
-}
